Implement event-driven route and error middleware registration
Event-based route registration is adopted in order to ensure
that the `TheMovieDB` instance imported in various controller
files is only initialized after a connection to the database
has been made. This is required because the `.initializeDatabase()`
method of the `TheMovieDB` class which may be called at construction
time(depending on the value of the `CONTENT_SOURCE` config variable)
attempts to make calls to the database.

Emitting appropriate events at different stages of the route
registration process, like just after a database connection has
been made, ensures the various routers are registered as intended.

Event-based error middleware registration is also adopted to ensure
that error-handling middleware are registered after all feature
routers.

The changes on a more atomic level:
	- move initialization code for `TheMovieDB` class to
	  `makeDatabaseFunction` in core/db.js
	- add event-related files to core directory
	- replace 'app.use()' with event-based 'registerFeatureRoute()'
	  and 'registerErrorMiddleware' in app.js(currently app-fb.js)
	- modify 'startServer()' to allow event-based invocation of
	  'app.listen()'

NOTE:
	An alternative solution would have been to 'await' the
	`makeDatabaseConnection` function before running code in any
	controller files (importing route.js files does this implicitly).
	This, however, would have resulted in less readable code or the
	use of ES Modules for top-level 'await' support.